% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobolib.R
\docType{package}
\name{sobolib}
\alias{sobolib}
\title{Sobo's Personal Package}
\description{
This package contains style templates and functions that I use often. If I use it more than a few times, I add it to this package!
}
\author{
Dennis Sobolewski \email{dennis.sobolewski@gmail.com}
}
