% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_binom.R
\name{summarise_binom}
\alias{summarise_binom}
\title{Calculates basic statistics for 0/1 binomial data}
\usage{
summarise_binom(tbl, col, ...)
}
\arguments{
\item{tbl}{A tibble of data.}

\item{col}{A binomial column in 0/1 or TRUE/FALSE format.}

\item{...}{Categorical columns that you wish .to group the analysis by}
}
\description{
Calculates basic statistics for 0/1 binomial data
}
